Cherry-pick 261785@main (8684905426f0). https://bugs.webkit.org/show_bug.cgi?id=254018
    Expose RealtimeMediaSource fitnessScore as a double
    https://bugs.webkit.org/show_bug.cgi?id=254018
    rdar://problem/106801067

    Reviewed by Eric Carlson.

    We should not silently cast double into unsigned, especially since we are doing comparison on the values.

    * Source/WebCore/platform/mediastream/RealtimeMediaSource.h:
    * Source/WebCore/platform/mediastream/RealtimeMediaSourceCenter.h:

    Canonical link: https://commits.webkit.org/261785@main

+++ Source/WebCore/platform/mediastream/RealtimeMediaSource.h
@@ -154,7 +154,7 @@ public:
     const AtomString& name() const { return m_name; }
     void setName(const AtomString& name) { m_name = name; }
 
-    unsigned fitnessScore() const { return m_fitnessScore; }
+    double fitnessScore() const { return m_fitnessScore; }
 
     WEBCORE_EXPORT void addObserver(Observer&);
     WEBCORE_EXPORT void removeObserver(Observer&);
